+++ compuglobal/models/episode.py
@@ -30,6 +30,10 @@ class EpisodeMetadata(BaseCompuGlobalModel):
         The original air date of the episode (yyyy-mm-dd)
     wiki_link : str
         The wikipedia link for the episode
+    video_width : int
+        The width of the video in pixels
+    video_height : int
+        The height of the video in pixels
 
     """
 
@@ -42,6 +46,8 @@ class EpisodeMetadata(BaseCompuGlobalModel):
     writer: str = Field(alias="Writer")
     original_air_date: str = Field(alias="OriginalAirDate")
     wiki_link: str = Field(alias="WikiLink")
+    video_width: int = Field(alias="VideoWidth", ge=0, default=480)
+    video_height: int = Field(alias="VideoHeight", ge=0, default=360)
 
 
 class EpisodeSummary(BaseCompuGlobalModel):