+++ compuglobal/models/frame.py
@@ -17,12 +17,18 @@ class Frame(BaseCompuGlobalModel):
         The episode key (S01E01)
     timestamp : int
         The timestamp of the frame
+    video_width : int
+        The width of the video in pixels
+    video_height : int
+        The height of the video in pixels
 
     """
 
     id: int = Field(alias="Id")
     key: str = Field(alias="Episode")
     timestamp: int = Field(alias="Timestamp", ge=0)
+    video_width: int = Field(alias="VideoWidth", ge=0, default=480)
+    video_height: int = Field(alias="VideoHeight", ge=0, default=360)
 
     @property
     def timecode(self) -> str: