+++ docs/changelog.rst
@@ -2,6 +2,14 @@
 
 Changelog
 =========
+
+Upcoming Changes
+----------------
+
+Miscellaneous
+~~~~~~~~~~~~~
+- Reorder fonts in :meth:`FontFamily.universal_fonts` and :attr:`CompuGlobalAPIConfig.allowed_fonts` to match the order used by the API.
+
 0.4.2
 -----